Improve text grammar
Both "at a high level" and "on a high level" are correct phrases in English, but they are used in slightly different contexts. "At a high level" is commonly used to describe a general or overview perspective, while "on a high level" can be used to indicate a position or location on a high surface or level.